Add: models for news rss.

+++ app/src/main/java/pl/mirror/black/blackmirror/model/weather/Main.java
@@ -6,9 +6,9 @@ public class Main {
 
     public Double temp;
 
-    public Integer pressure;
+    public Double pressure;
 
-    public Integer humidity;
+    public Double humidity;
 
     @SerializedName("temp_min")
     public Double tempMin;
